feat: add domain and presenter

+++ api/domain/repository/story.go
@@ -3,6 +3,6 @@ package repository
 import "github.com/keinuma/go-graphql/api/domain/entity"
 
 type StoryRepository interface {
-	GetStories(limit, offset int64) (*entity.Stories, error)
+	GetStories(limit, offset int) (*entity.Stories, error)
 	CreateStory(story entity.Story) (*entity.Story, error)
 }
